+++ packages/runtime/drqp_control/src/JointsStateToPose.cpp
@@ -18,56 +18,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <cstdint>
 #include <exception>
-#include <memory>
-#include <string>
-#include <unordered_map>
-#include <algorithm>
 
-#include "drqp_control/DrQp.h"
+#include "drqp_control/JointServoMappings.h"
 
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/joint_state.hpp>
 
 #include <drqp_interfaces/msg/multi_sync_position_command.hpp>
 #include <drqp_interfaces/msg/multi_async_position_command.hpp>
-
-struct ServoParams
-{
-  uint8_t id;
-  double ratio = 1.;
-};
-struct JointParams
-{
-  std::string jointName;
-  double ratio = 1.;
-};
-
-static const auto kJointToServoId = []() {
-  std::unordered_map<std::string, ServoParams> jointToServoId;
-  std::unordered_map<uint8_t, JointParams> servoIdToJoint;
-
-  const std::string kRight = "right";
-  const std::array<const char*, kServosPerLeg> kJointNames = {"_coxa", "_femur", "_tibia"};
-  for (const auto& leg : kAllLegServoIds) {
-    int jointNameIndex = 0;
-    for (const uint8_t servoId : leg) {
-      const std::string legName = legNameForServo(servoId);
-      std::string jointName = "dr_qp/" + legName + kJointNames[jointNameIndex];
-
-      const bool isRight =
-        std::find_end(legName.begin(), legName.end(), kRight.begin(), kRight.end()) !=
-        legName.end();
-      const bool isCoxa = jointNameIndex == 0;
-      double ratio = (isRight && !isCoxa) ? -1. : 1.;
-      jointToServoId[jointName] = {servoId, ratio};
-      servoIdToJoint[servoId] = {jointName, ratio};
-      ++jointNameIndex;
-    }
-  }
-  return jointToServoId;
-}();
 
 class JointsStateToPoseNode : public rclcpp::Node
 {
