+++ packages/runtime/drqp_control/src/PoseToJointsState.cpp
@@ -26,41 +26,13 @@
 #include <algorithm>
 
 #include "drqp_control/DrQp.h"
+#include "drqp_control/JointServoMappings.h"
 
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/joint_state.hpp>
 
 #include <drqp_interfaces/msg/multi_sync_position_command.hpp>
 #include <drqp_interfaces/msg/multi_async_position_command.hpp>
-
-struct Params
-{
-  std::string name;
-  double ratio = 1.;
-};
-
-static const auto kServoIdToJoint = []() {
-  std::unordered_map<uint8_t, Params> result;
-
-  const std::string kRight = "right";
-  const std::array<const char*, kServosPerLeg> kJointNames = {"_coxa", "_femur", "_tibia"};
-  for (const auto& leg : kAllLegServoIds) {
-    int jointNameIndex = 0;
-    for (const int servoId : leg) {
-      const std::string legName = legNameForServo(servoId);
-      std::string jointName = "dr_qp/" + legName + kJointNames[jointNameIndex];
-
-      const bool isRight =
-        std::find_end(legName.begin(), legName.end(), kRight.begin(), kRight.end()) !=
-        legName.end();
-      const bool isCoxa = jointNameIndex == 0;
-      double ratio = (isRight && !isCoxa) ? -1. : 1.;
-      result[servoId] = {jointName, ratio};
-      ++jointNameIndex;
-    }
-  }
-  return result;
-}();
 
 class PoseToJointState : public rclcpp::Node
 {
@@ -121,10 +93,10 @@ private:
       return;
     }
 
-    const Params params = kServoIdToJoint.at(servoId);
+    const JointParams params = kServoIdToJoint.at(servoId);
     const double positionInRadians = params.ratio * positionToRadians(position);
 
-    jointState.name.push_back(params.name);
+    jointState.name.push_back(params.jointName);
     jointState.position.push_back(positionInRadians);
   }
